+++ Import.ps1
@@ -1,0 +1,1 @@
+Import-Module ./src/Evergreen.psd1 -Verbose -Force