+++ ci/Deploy.ps1
@@ -29,7 +29,7 @@ Else {
         $projectRoot = Resolve-Path -Path (((Get-Item (Split-Path -Parent -Path $MyInvocation.MyCommand.Definition)).Parent).FullName)
         $module = Split-Path -Path $projectRoot -Leaf
     }
-    $moduleParent = Join-Path -Path $projectRoot -ChildPath $module
+    $moduleParent = Join-Path -Path $projectRoot -ChildPath $source
     $manifestPath = Join-Path -Path $moduleParent -ChildPath "$module.psd1"
     #$modulePath = Join-Path -Path $moduleParent -ChildPath "$module.psm1"